+++ service/src/main.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code)]
+#![feature(try_blocks)]
 
 use crate::storage_impl::create_tables;
 use marine_rs_sdk::module_manifest;
@@ -7,11 +8,15 @@ use marine_rs_sdk::WasmLoggerBuilder;
 module_manifest!();
 
 mod dto;
+mod error;
+mod misc;
 mod results;
 mod service_api;
-mod service_impl;
 mod storage_impl;
 mod tests;
+
+pub static TRUSTED_TIMESTAMP_SERVICE_ID: &str = "peer";
+pub static TRUSTED_TIMESTAMP_FUNCTION_NAME: &str = "timestamp_sec";
 
 pub fn main() {
     WasmLoggerBuilder::new()
